requirements.txt
Pillow==10.2.0
fonttools==4.49.0